requirements.txt
mandrill==1.0.57
mock==3.0.5
Django==2.2.1
djangorestframework==3.10.1
